+++ admin/tabs/general/GeneralTab.js
@@ -6,7 +6,7 @@
  */
 import {configPanel} from '@xh/hoist/admin/tabs/general/config/ConfigPanel';
 import {tabContainer} from '@xh/hoist/cmp/tab';
-import {hoistCmp} from '@xh/hoist/core';
+import {XH, hoistCmp} from '@xh/hoist/core';
 import {Icon} from '@xh/hoist/icon';
 import {aboutPanel} from './about/AboutPanel';
 import {alertBannerPanel} from './alertBanner/AlertBannerPanel';
@@ -21,9 +21,14 @@ export const generalTab = hoistCmp.factory(
             tabs: [
                 {id: 'about', icon: Icon.info(), content: aboutPanel},
                 {id: 'config', icon: Icon.settings(), content: configPanel},
-                {id: 'users', icon: Icon.users(), content: userPanel},
+                {id: 'users', icon: Icon.users(), content: userPanel, omit: hideUsersTab()},
                 {id: 'alertBanner', icon: Icon.bullhorn(), content: alertBannerPanel}
             ]
         }
     })
 );
+
+const hideUsersTab = () => {
+    const conf = XH.getConf('xhAdminAppConfig', {});
+    return conf['hideUsersTab'];
+};